atet_fixture: Move Fixture class out of arcyd-tester
Extract _Fixture class out of arcyd-tester into a new file
atet_fixture.py. Not adding unit tests explicitly for this class as all
functions are exercised in arcyd-tester, especially those for the happy
path and unit tests would simply duplicate that.

Test plan:
$ ./precommit.sh
